gate authorize in controller

+++ app/Http/Controllers/CommentController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Post;
 use App\Models\Comment;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 use App\Http\Requests\CommentRequest;
 
 class CommentController extends Controller
@@ -15,12 +16,18 @@ class CommentController extends Controller
     }
 
     function update(CommentRequest $request,Comment $comment){
-        $comment->update($request->all());
-        return back()->with('success', 'Comment Update Success');
+        if(Gate::allows('comment-update', $comment)){
+            $comment->update($request->all());
+            return back()->with('success', 'Comment Update Success');
+        }
+        return abort(403, 'Unauthorized action.');
     }
 
     function destroy(Comment $comment){
-        $comment->delete();
-        return back()->with('success', 'Comment Delete Success');
+        if(Gate::allows('comment-destroy',[$comment,$comment->post])){
+            $comment->delete();
+            return back()->with('success', 'Comment Delete Success');
+        }
+        return abort(403, 'Unauthorized action.');
     }
 }
